# Signal: create platform-specific generic conversion methods to avoid boxing doubles