--- a/src/ScottPlot4/ScottPlot/Plottable/MinMaxSearchStrategies/LinearFastDoubleMinMaxSearchStrategy.cs
+++ b/src/ScottPlot4/ScottPlot/Plottable/MinMaxSearchStrategies/LinearFastDoubleMinMaxSearchStrategy.cs
@@ -41,7 +41,7 @@ namespace ScottPlot.MinMaxSearchStrategies
         {
             if (sourceArrayDouble != null)
                 return sourceArrayDouble[index];
-            return Convert.ToDouble(SourceArray[index]);
+            return NumericConversion.GenericToDouble(ref SourceArray[index]);
         }
     }
 }