**Purpose:**
Generic implementations such as `SegmentedTree<T>` fall back to `Convert.ToDouble` requiring the values to be boxed.
We can implement a fast path for well known and frequently used types (ex: doubles, floats).

Take this example: 
```cs
double M<T>(T value)
{
    // same as typeof(T) == typeof(double)
    if (value is double d) 
        return d;
    else if (value is float f)
        return f;
    else
        return Convert.ToDouble((object)value); // implicit cast to object
}
```
is jitted to:
```cs
double M(double value) => value;
double M(float value) => (double)value;
double M<T>(T value) => Convert.ToDouble((object)value); // implicit cast to object
```

Notes:
- `Convert.ToDouble` is overkill for a simple cast
- `Convert.ToDouble` requires boxing the value when the type is not known
- `#if NET` to avoid regression on old frameworks
- We already optimize some path using `System.Linq.Expressions.Expression` but i prefer this solution

Repro:
```cs
var rnd = new Random(123);
var x = 0d;
var ys = ScottPlot.DataGen.RandomWalk(rnd, 10_000_000);
var xs = Enumerable.Range(0, ys.Length).Select(_ => x += rnd.NextDouble() * 1000).ToArray();

var plot = new ScottPlot.Plot(1000, 1000);
plot.AddSignalXYConst(xs, ys);
plot.Render(lowQuality: true);
```

Result:
||Type|Alloc|Time|Notes
|-|-|-|-|-|
|Now|double|19.964.478|480ms|10M validation + (10M * 1 refresh)|
|After|double|7.154|415ms||