@swharden 2022-06-30
> This is an incomplete implementation, i want some feedback before proceeding.

Hi @AbeniMatteo, thanks for looking into this!

We can get performance improvements by adding version-specific optimized code blocks that optimally handle generic values. Enhanced performance is achieved at the cost of code that gets more and more complex and harder to maintain (with duplicated logic for different platforms and data types), so these are the two sides I'm trying to balance between.

The small `double`-specific optimizations so far in this PR seem worth it, and I'd be happy to merge something like this in when you're done.

> We can implement a fast path for well known and frequently used types

Rather than go down that road now, maybe we can get a huge win for a small code change if we use generic math in .NET 7? I've been waiting for this feature forever, and it solves this exact problem. Between `Span<T>` and this, signal plots could become very performant with simple code that doesn't duplicate logic

https://devblogs.microsoft.com/dotnet/dotnet-7-generic-math/

What do you think?