@swharden 2022-06-30
Sounds good! 

My only suggestion is to not throw it in `Tools` because big catch-all static libraries that collect random stuff are a bit of an anti-pattern and I wish I hand't created that module.

Maybe a small static class with a better name would be more expressive and portable. Maybe something like ´NumericConversion.DoubeToGeneric()´?

Thanks again for your work on this!