@AbeniMatteo 2022-06-30
The implementation is complete. 
I still have to test and profile the results.

___

I also tried generic math, you can play with it [here](https://github.com/AbeniMatteo/ScottPlot/tree/ab-generic-math).
Some notes:
- Conversion from `T` to `double` still requires a `Convert.ToDouble` or this PR
- Supporting mixed compilation for >NET7 and <NET7 isn't trivial
- I'm currently using `INumber `as generic restraint even tho we need less operations (ex: `IAdditionOperators`+`IComparisonOperators`)
- `T.Max()` and `T.Min()` are only implemented on `INumber`
- Support for TimeSpan and DateTime was dropped
https://github.com/dotnet/runtime/pull/69331
- A solution the previous point is that it should be possible to write a simple struct wrapper around non INumber-types the user want to use
```cs
struct Wrapper : IAdditionOperators<Wrapper, Wrapper, Wrapper>, IConvertible
{
    public Wrapper(TimeSpan span) { }

    public static Wrapper op_CheckedAddition(Wrapper left, Wrapper right) { throw null; }
    public static Wrapper operator +(Wrapper left, Wrapper right) { throw null; }
    public double ToDouble(IFormatProvider provider) { throw null; }
}
```