@bclehmann 2022-07-01
Don't take this as a suggestion for this PR, rather it's a request for input on something I've been considering for a while:

It might be worth creating a benchmarks project, just so that further performance work can rely on less ad hoc testing?

I still expect that most performance-related PRs would require additional benchmarks to be written for validation, it would just streamline the process. For example, we could hook it up to [BenchmarkDotNet ](https://github.com/dotnet/BenchmarkDotNet) and not have to write our own throwaway code to track timing and allocations (or rely on an external profiler, which is less convenient and reproducible, at least for me).

There is also the option to make benchmarks and certain minimum performance characteristics part of CI. This is probably not even worth considering until we have a large number of benchmarks written (and perhaps not even then), but I thought I'd bring it up.