@AbeniMatteo 2022-07-01
Micro benchs results are really good:
- Gains for `net5` and `net6` are massive
- `netcore2` shows no improvements (i'm using `#if NETCOREAPP`), i think this optimization is available from `netcore3`
- no regression on `.NET framework` for small values like `float` and `double`
- small regression on <`net20` for big types like `decimal`
```cs
public double New(T value)
{
    var a = Utils.ConvertGenericToDouble(value);
    a *= 999;
    a -= 5.6;
    return a;
}

public double Old(T value)
{
    var a = Convert.ToDouble(value);
    a *= 999;
    a -= 5.6;
    return a;
}
```

|Type    | Method |       Jit |              Runtime |     Mean | Code Size |
|-------|------- |---------- |--------------------- |---------:|----------:|
|double|    New |    RyuJit |             .NET 5.0 |  0.2427 ns |      20 B |
|double|    Old |    RyuJit |             .NET 5.0 |  9.5874 ns |     136 B |
|double|    New |    RyuJit |             .NET 6.0 |  0.0041 ns |      20 B |
|double|    Old |    RyuJit |             .NET 6.0 | 10.2222 ns |     136 B |
|double|    New |    RyuJit |        .NET Core 2.0 |  9.0733 ns |     420 B |
|double|    Old |    RyuJit |        .NET Core 2.0 | 10.2520 ns |     123 B |
|double|    New |    RyuJit | .NET Framework 4.7.2 |  9.9730 ns |     135 B |
|double|    Old |    RyuJit | .NET Framework 4.7.2 |  9.9856 ns |     135 B |
|double|    New | LegacyJit | .NET Framework 4.7.2 |  9.9936 ns |     135 B |
|double|    Old | LegacyJit | .NET Framework 4.7.2 | 10.1067 ns |     135 B |
|-||||||||
|float|    New |    RyuJit |             .NET 5.0 |  0.4692 ns |      24 B |
|float|    Old |    RyuJit |             .NET 5.0 | 10.0318 ns |     136 B |
|float|    New |    RyuJit |             .NET 6.0 |  0.4512 ns |      24 B |
|float|    Old |    RyuJit |             .NET 6.0 | 10.5184 ns |     136 B |
|float|    New |    RyuJit |        .NET Core 2.0 | 12.4097 ns |     420 B |
|float|    Old |    RyuJit |        .NET Core 2.0 | 10.1572 ns |     123 B |
|float|    New |    RyuJit | .NET Framework 4.7.2 |  9.9812 ns |     135 B |
|float|    Old |    RyuJit | .NET Framework 4.7.2 |  9.9965 ns |     135 B |
|float|    New | LegacyJit | .NET Framework 4.7.2 | 10.0919 ns |     135 B |
|float|    Old | LegacyJit | .NET Framework 4.7.2 | 10.2973 ns |     135 B |
|-||||||||
|decimal|    New |    RyuJit |             .NET 5.0 | 13.96 ns |    147 B |
|decimal|    Old |    RyuJit |             .NET 5.0 | 14.42 ns |    133 B |
|decimal|    New |    RyuJit |             .NET 6.0 | 13.65 ns |    147 B |
|decimal|    Old |    RyuJit |             .NET 6.0 | 14.92 ns |    133 B |
|decimal|    New |    RyuJit |        .NET Core 2.0 | 25.17 ns |    462 B |
|decimal|    Old |    RyuJit |        .NET Core 2.0 | 17.91 ns |    123 B |
|decimal|    New |    RyuJit | .NET Framework 4.7.2 | 16.87 ns |    143 B |
|decimal|    Old |    RyuJit | .NET Framework 4.7.2 | 15.90 ns |    134 B |
|decimal|    New | LegacyJit | .NET Framework 4.7.2 | 17.02 ns |    143 B |
|decimal|    Old | LegacyJit | .NET Framework 4.7.2 | 15.47 ns |    134 B |